Change behavior of DeferredResult to resolve to false on error from STS. Make MessageQueue not block while it drains its queue after enqueing. and some PR fixes.
