+++ javatests/arcs/android/storage/service/ParcelableProxyMessageChannelTest.kt
@@ -73,18 +73,11 @@ class ParcelableProxyMessageChannelTest {
 
     @Test
     fun proxyMessages_whenHandledResultIsFalse_sendExceptionToCallback() = runBlocking {
-        try {
-            supervisorScope {
-                val channel = ParcelableProxyMessageChannel(coroutineContext)
-                val deferredResult = DeferredResult(coroutineContext)
-                channel.onProxyMessage(makeMessage(0), deferredResult)
-                channel.openSubscription().receive().result.complete(false)
-                deferredResult.await()
-                fail("Shouldn't get here.")
-            }
-        } catch (e: Exception) {
-            assertThat(e.message).contains("returned false as result")
-        }
+        val channel = ParcelableProxyMessageChannel(coroutineContext)
+        val deferredResult = DeferredResult(coroutineContext)
+        channel.onProxyMessage(makeMessage(0), deferredResult)
+        channel.openSubscription().receive().result.complete(false)
+        assertThat(deferredResult.await()).isFalse()
     }
 
     @Test
@@ -97,8 +90,8 @@ class ParcelableProxyMessageChannelTest {
                 channel.openSubscription().receive().result
                     .completeExceptionally(CrdtException("Uh Oh!"))
                 deferredResult.await()
-                fail("Shouldn't get here.")
             }
+            Unit
         } catch (e: Exception) {
             assertThat(e.message).contains("Uh Oh!")
         }
