+++ javatests/arcs/core/storage/util/ProxyCallbackManagerTest.kt
@@ -56,9 +56,6 @@ class ProxyCallbackManagerTest {
             }
         }.joinAll()
 
-        // Now check that the nextCallbackToken is correct (should be 201 if we didn't erroneously
-        // attempt to increment outside of a lock.
-        assertThat(manager.nextCallbackToken.value).isEqualTo(201)
         assertThat(manager.callbacks.size).isEqualTo(200)
     }
